Further simplify kodion.ui module structure
- Move AbstractProgressDialog into kodion.ui.abstract_context_ui
- Remove concrete XbmcProgressDialog class
- Use existing XbmcContextUI.create_progress_dialog class factory method to create instances of AbstractProgressDialog directly
